[feature] implement chain of responsibilities pattern
