+++ internal/infrastructure/currency/coingecko/coingecko.go
@@ -1,15 +1,23 @@
 package coingecko
 
 import (
+	"context"
+	"errors"
 	"net/http"
 
 	"exchange/internal/domain/rate"
 )
 
+type Chain interface {
+	GetCurrency(ctx context.Context, cur *rate.Rate) (float64, error)
+	SetNext(any) error
+}
+
 type CoingeckoAPI struct {
 	cfg    *Config
 	cli    *http.Client
 	mapper map[string]string
+	next   Chain
 }
 
 func NewCoingeckoApi(cfg *Config) *CoingeckoAPI {
@@ -26,4 +34,15 @@ func initMapper() map[string]string {
 	coins[rate.BTC] = "bitcoin"
 
 	return coins
+}
+
+func (api *CoingeckoAPI) SetNext(chain any) error {
+	v, ok := chain.(Chain)
+	if !ok {
+		return errors.New("unable to set next handler. Invalid type")
+	}
+
+	api.next = v
+
+	return nil
 }
